Add a special-case for when someone tries to read from hardware port 513 on Window NT.
Reading from hardware port 513 is how you read the joystick buttons in Logo.
On NT, you cannot read directly from the port, so I used the Win32 API for
getting the joystick button states and then reassemble something that is close
to what it at that hardware port.  It's not an exact replica, but I think
it will suffice for how people use it.

Maybe I should add a new Logo command for getting the button state.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@92 9d0050d7-00b0-40f9-871c-8e02cbd5794f