Add lots of ASSERTs to catch memory corruption of the turtle array.
Fix the regression introduced in the previous commit.  The problem was
that a local variable shadowed a global variable and I didn't realize this.
When I refactored the code, it was changed to use the global variable,
which change the meaning of the code.

Replce the OldBrush global variable with stack variables.

Miscellaneous formatting changes.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@111 9d0050d7-00b0-40f9-871c-8e02cbd5794f