Implement feature request #1365734; FULLSCREEN, TEXTSCREEN, and SPLITSCREEN now undock and dock the commander window.  This was primarily done to support "real" applications--ones that don't allow Logo input, such as arcade-style games.
The implementation is a big hack.  I can't just undock and redock the window, because the splitter widgit that I use destroys the window when it's removed.  As a result, I have to create a new commander window, copy the contents from the old one, and delete the old one.

Adding an undocked commander required changing a lot of other code that had been assuming the commander was always docked.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@310 9d0050d7-00b0-40f9-871c-8e02cbd5794f