Fix bug #1372735.
Replace a problematic call to Editor->GetWindowText() with Editor->GetSubText().  GetWindowText() has been the root of two different bugs that cause FMSLogo to fail to save the contents of the editor.

The latest bug only showed up on Win95 and it looks like GetWindowText() returned a 16-bit signed value that was cast to an int.  For example, a file of about 60000 bytes returned about -5600 bytes read.

I have not encounted any problems with GetSubText() yet.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@402 9d0050d7-00b0-40f9-871c-8e02cbd5794f