Split setpos_helper into three functions: 1) setpos_helper() uses uses NODE*. 2) setpos_helper_2d() uses FLONUM and assumes 2D mode. 3) setpos_helper_3d() uses FLONUM and assumed 3D mode.
Rewrite lellipsearc() to use the *_2d and *_3d routines.  The result is a 5% speed gain and fewer chances for allocations to fail.  I am disapponted that the speed gain is so small, but it's better than it was (and I don't feel like I sacrified anything in code quality to get it).


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@430 9d0050d7-00b0-40f9-871c-8e02cbd5794f