Fix bug #1372489; double-clicking a .lgo file now starts a new FMSLogo process.  Running FMSLogo without a -L command-line parameter will just give focus to any instance of FMSLogo that is already running.
This keeps the (good) behavior of when a child double-clicks on the explorer shortcut for FMSLogo many times, but ends up with only one empty instance of FMSLogo.   This fixes an annoyance that double-clicking on a .LGO file just gave focus to FMSLogo, instead of running the program.

This is not an ideal solution, but I think it's good enough, considering how complicated the better solutions are.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@735 9d0050d7-00b0-40f9-871c-8e02cbd5794f