Fix bug #1422237; AM.lgo can now be opened, saved, and re-run in Logo.
The problem was that the code was burying some of its critical functions.  Based on the readme, I believe this was to thwart users from figuring out how to put profanity into the code.  My intent in releasing this code was that anyone could easily extend it, so I'm more concered about someone getting frustrated trying to modify AM.lgo than someone preverting their private copy of it.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@736 9d0050d7-00b0-40f9-871c-8e02cbd5794f