Remove setint() and setfloat().  Rewrite getint() and getfloat() to be inline functions so that they can't accidently be used to modify the value of a node.
This helps enforce the rule that, once an int and a float are created, their value never changes.  At some point, I may use this fact to pre-create some common numeric nodes so that they don't have to be created at runtime.  This was a common LISP optimization back in the day and it may still be relevant today.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@793 9d0050d7-00b0-40f9-871c-8e02cbd5794f