Refactor the tests such that each test loads "testlib.lgo, instead of assuming that it is loaded by the test harness.  This allows individual tests to be loaded and run on their own.  This is useful for reproducing and fixing bugs now that the whole test suite takes quite a while to run.
To do this, I had protect against the case where testlib.lgo is loaded while a test is running.  I also had to rename arc.lgo to curves.lgo so that it wouldn't get loaded when ARC was executed.  I also had to remove a misfeature in LOAD that would prompt you if the workspace was dirty (bug #1392437)


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@798 9d0050d7-00b0-40f9-871c-8e02cbd5794f