Fix bug #1397549, ARCSIN and friends now throw an error on invalid input.
This also addresses bug #1395743 and bug #1395742, which complained about the dialog box shown in domain errors.  Now no dialog box is shown.  This is consistent with my and Brian Harvey's belief that excessive dialog boxes ruins the language by coupling it with non-language elements.

This was accomplished by overriding the default Borland math error handling, which displays a dialog box whenever a C trig function is given bad input.  This is not Logo-like--it should throw a Logo error instead of halting the program. By doing a long jump, the calling code will handle it equivalent to how they handle floating point exceptions.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@820 9d0050d7-00b0-40f9-871c-8e02cbd5794f