Fix bug #1432651; having a filename in the current working directory that matches the name of a library procedure no longer crashes.
This turned out to be a problem with reference counting.  For some reason the code that works was commented out and the buggy code was used in its place.  I don't know why the working code was switched out, but switching it back in fixes the crash without causing a memory leak, so I'm
pretty sure it's the right change.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@917 9d0050d7-00b0-40f9-871c-8e02cbd5794f