Changes to improve the debug-ability of FMSLogo:
Change getobject from a macro to an inline function that has an assert failure on bad input.  Change cddr and friends to use car() and cdr(), which also assert on bad input.  This change also ensures that we are not assigning directly to a cddr() because the return value is not an lvalue.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@918 9d0050d7-00b0-40f9-871c-8e02cbd5794f