Several improvements to make the tests more effective:
In RUNEXPECTERRORTEST, when the instructions throw the wrong error code, print out the error code that was expected, in addition to what was thrown.

In RUNBADINPUTTEST, RUNEXPECTERRORTEST, and RUNPERFORMANCETEST, use RUNRESULT, instead of RUN so that these test don't throw an error when given an instruction that doesn't output.  This found several test bugs that assumed something throw an error when really the command worked fine, but RUNBADINTPUTTEST threw an error because it expected the instructions to return something.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@940 9d0050d7-00b0-40f9-871c-8e02cbd5794f