Add some missing "const" qualifiers to the real_print_node routine.  Before, I couldn't use const because the code could modify the node to print it.  I did this because my debugheap would crash when I printed some leaked nodes.
To do this, I ended up duplicating code that is in the node-to-string conversion routine.  I think this a little code duplication is better than removing the node dumping code in in the debugheap.  It also has the side-benefit of not allocating anything to print a node.

This change also includes some cosmetic changes, such as adding missing curly braces and moving stack varible declarations to the line on which they are first used.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@952 9d0050d7-00b0-40f9-871c-8e02cbd5794f