Fix bug #1448142; COPYDEF'ing STOP no longer crashes
The fix was to assign special priorities to STOP, OUTPUT, OP, and .MAYBEOUTPUT so that they can be recognized, not by name, but by priority.  I don't know if this is the right way to do it, but it's what UCBLogo does, so it's what FMSLogo will end up doing if it ever updates to the latest language engine.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@957 9d0050d7-00b0-40f9-871c-8e02cbd5794f