Save the program in FMSLogo, which alphabetises the functions by name, making it too difficult to see what has changed by doing a file diff.
Rescope most variables as LOCAL so that they don't show up at the bottom of EDALL.

Move the definition of the game board template into a procedure so that the whitespace is preserved.  This makes it easier for children to modify the game board.

Fix some bugs in the array initializer.  I quoted the string, which was already quoted by the curly braces.  Among other things, this cause the printout to say "Blue Wins, instead of just Blue Wins.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@992 9d0050d7-00b0-40f9-871c-8e02cbd5794f