Implemented Feature Request #1434420: Remove the text limit on the commander's history box.
The fix was surprisingly invasive, as it required that I change the commander box from a TEdit to a TRichEdit.  I had to do several hacks to ensure that the TRichEdit control acted like the previous TEdit, even though the two classes are very difference.

One side-benefit of this is that it will make it easier to color-code commands differently from responses.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1167 9d0050d7-00b0-40f9-871c-8e02cbd5794f