Fix bug #1395736; CTRL+A now works in the commander.
The fix required adding an accelerator for CTRL+A to EV_EDITSELECTALL to the main window's accelerator table.  This ended up blocking the accelerator translation for the editor, so I had to add some plumbing to ensure that the message was delegated all the way to the editor's TRichEdit.

I also added a "Select All" option to the editor's menu.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1171 9d0050d7-00b0-40f9-871c-8e02cbd5794f