Fix a per-widget memory leak of the dialogthing* pointer.
I was unable to fix the memory leak with a simple change, so I took the oppertunity to refactor the dialogthing into the slink class.  I also moved the "key" field that was in all of the widget classes into the dialogthing class, which is not acting as a sort of base class.  The end result is that I was able to remove a good deal of repeated code, which is a good thing.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1182 9d0050d7-00b0-40f9-871c-8e02cbd5794f