Refactor to eliminate some duplicated in all of the lcreate* functions.
Now that windows and dialogs use the same base class, there was no need to have separate (but very similar) code paths when adding to a window versus when adding to a dialog.  The new code treats adding to a window or a dialog the same (although it preserves the difference between the two in the "type" field).


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1199 9d0050d7-00b0-40f9-871c-8e02cbd5794f