Fix bug #1500217; LISTBOXGETSELECT now outputs [] when nothing is selected (instead of outputting nothing).  Outputting nothing was a problem because trying to use the output could potentially throw a "does not output to" error.
While there is theoretical benefit to being able to distinguish between when nothing is selected and when the empty string is selected, I decided to output [] when nothing is selected instead of throwing an error because this is a very special case that doesn't ever happen in practice and I didn't want to add a new error type for such an unlikely condition.  This is also consistent with how COMBOBOXGETTEXT behaves.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1226 9d0050d7-00b0-40f9-871c-8e02cbd5794f