Fix bug #1367386: the network code no longer goes into an infinite loop when sent data that does not contain NUL.
The root cause of the problem is that calling recv() will reset the AsyncSelect() state, so it will immediately notify us that there was still data in the queue if we don't remove it.  The fix is to always remove all data from the recv() queue and store it in a CCarryOverData class if there isn't a complete datagram.

I also made FMSLogo send up any partial packets when the sockets are closed.  This allows a small chance at interoperating with real TCP-based protocols that don't append a NUL to frame the packets.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1340 9d0050d7-00b0-40f9-871c-8e02cbd5794f