Fix bug #1455201: Running TO from STARTUP no longer hangs FMSLogo.
The problem was that to_helper() would try to reading from a file stream that was already closed.  This caused to_helper() to think it read a blank line and try to read some more data.

The fix is to restore the previous value of loadstream before calling TO.  The code should also probably error-out if ferror() returns non-zero, but I want to think about how to best integrate that into the code, as this should probably be done in many, many places.

I have also added a regression test for this bug.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1356 9d0050d7-00b0-40f9-871c-8e02cbd5794f