Remove an extra copy of the  network packets by just setting network_send_value and network_receive_value to the data in the callevent, instead of coping the data into a per-allocated buffer.
This probably leads to faster code, but that's not why I made the change.  The real motivation is that this eliminates the possibility of a race condition between receiving the last packet and shutting down/starting up the client/server.  My fear was that at some point, the code might try to copy data into a buffer that was freed.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1367 9d0050d7-00b0-40f9-871c-8e02cbd5794f