Fix two things which looked weird, although they may not have been bugs:
First, if we are given an unexpected asynchronous select message, we no longer show a meaningless message box to the programmer/user.  This is a bug in FMSLogo (or Windows) if it happens and there's nothing actionable for the Logo programmer.

Second, we no longer set the m_IsConnected flag or call the SendReady callback if accept() fails.  Doing so is wrong because we don't actually have a connection.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1390 9d0050d7-00b0-40f9-871c-8e02cbd5794f