Fix bug #1517326: ERRACT value that are anonymous functions no longer cause FMSLogo to crash.
The problem was that in some code paths, the ERRACT instuction list was getting deref'd within the evalutator.  For the non-erract case, this protected against by explicitly referencing the list input, but that code was skipped by the erract evaluator.  The fix is for evaluator() to always reference the "list" parameter upon entering.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1420 9d0050d7-00b0-40f9-871c-8e02cbd5794f