Fix bug #1458661; Macros that output nothing no longer crash FMSLogo.
Here is a description of the bug:
RUNRESULTcreate a continuation node.  Continuation nodes store a
number, not a NODE* in their "CAR" field (the ID of where
they should continue evaluating from).  Normally, this is
overwritten with NIL when the continuation is used.
However, if there's an error, the continuation is cleaned up
before the CAR field is cleared to NIL.  And there was a bug
in gc(), which is where all NODE* are cleaned up, that tried
to free the CAR of continuation nodes.  This resulted in an
access violation, as you obviously can't free() the number 11.

The fix is to only free the CDR of a continuation NODE* in gc().

This fix will be released in FMSLogo 6.11.0.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1543 9d0050d7-00b0-40f9-871c-8e02cbd5794f