Restructure the COPYDEF tests to look more like my other tests.
I also switched in some tests for bug #1454113 which no longer crash.  I added a new test that uses COPYDEF in a recursive, but not tail-recursive call, which still crashes due to bug #1454113.  This test case is currently commented-out.

Also, fix a bug in EVALTEST.TAILRECURSIVECOUNT that made it always output 0.  This bug led to two false-positives per test pass.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1573 9d0050d7-00b0-40f9-871c-8e02cbd5794f