Update the documentation based on changes to the error handling subsystem that were made in the last release.  CATCH "ERROR no longer unbinds "Erract" and you cannot shadow CATCH "ERROR by re-defining Erract.
Now that I have read how ERRACTused to behave, I'm not so confident that I did the right thing by changing the behavior.  Still, the documentation should match the current behavior.  If I later revert the changes to the code, then I will also revert the changes to the manual.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1576 9d0050d7-00b0-40f9-871c-8e02cbd5794f