Fix bug #1563318; This is a regression that caused FMSLogo to crash when running procedures with complex inputs that include chains of default values and rest inputs.
The bug is that the formal parameters of a proc node can be freed while the proc node is being used.  This bug can be traced back to a memory leak fix in numberp(), but I'm fairly certain that fix is correct.  There must be some other place where one too many references is consumed (perhaps the ref count is zero, then incremented, then decremented).  Anyway, I couldn't find this place, but I was able to fix the crash by adding a reference on the formals inputs while they are being used.  Since this bug was reported by a user, it's better to get a good fix out, rather than waiting until the perfect fix can be found.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1606 9d0050d7-00b0-40f9-871c-8e02cbd5794f