Replace the movement (FD, BK, FORWARD, and BACK) and turn (RT, RIGHT, LT, and LEFT) primitives with the game movement commands so that typing in a commands still counts as a move can still crash/win.
Fix a bug where you could move through a line if you moved through it near the endpoint.  Now a tolerance of 1e5 is allowed for floating point precision.

Fix the cosmetic bug where crashing in one move reported the crash as "1 moves".  Now it is correctly reported as "1 move".

Add commands that have the same name as the RESET and NEXT buttons, for convenience.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1673 9d0050d7-00b0-40f9-871c-8e02cbd5794f