Fix bug #1584198; Commander selection no longer disappears when FMSLogo loses focus.
This bug can be traced back to when the commander's history box was made into a rich edit control.  At the time, I was trying to mimic MSWLogo's exact behavior and I saw that the selection disappeared when it lost focus.  What actually happened is that the highlighting disappeared, but the selection remained.  My fix is to clear the selection when the edit control gains focus, instead of when the history control loses it, which has a more natural feel.  I experimented with hiding/showing the selection, but I couldn't find something that felt natrual for the end user.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1717 9d0050d7-00b0-40f9-871c-8e02cbd5794f