Move the network test to the end of the test pass because running SLOWDRAWTEST after it causes problems (somehow, the call to WRITER in REPORTTESTINFO gets corrupted, halting the test pass).
This is a work-around for a bug.  I don't know if I recently added this to FMSLogo, or if the tests just recently became complicated to expose a long-standing bug.  Either way, I will file this before the next release, I just don't have time to look into this bug at the momment.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1739 9d0050d7-00b0-40f9-871c-8e02cbd5794f