Fix bug #1599367; SPLITSCREEN no longer moves the X/Y scrollbar position.
I refactored the code from the ZOOM scrollbar fix into TScreenWindow so that it could be used for this bug, as well.  It looks like this also fixes a related bug where the scroll positions drifted a little when the window was resized.  Based on the number of times the old code referenced MainWindowx->ScreenWindow, I think TScreenWindow is the right place for the code.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1749 9d0050d7-00b0-40f9-871c-8e02cbd5794f