Port the "deepend_proc_name" global variable from UCBLogo.  This lets the "END inside multi-line instruction" error message additionally display the name of the procedure that has the run-away line.
This is a very useful feature when writing long programs.  I expect that this will save me lots of time when developing my tests, because I often forget to close a "[" or "(" and have a hard time tracking down where the unclosed brace is.  This will significantly narrow my search.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1756 9d0050d7-00b0-40f9-871c-8e02cbd5794f