Internationalize the remaining string literals that are part of the UI by moving them into a localizedstrings-*.h.
I also took this opportunity to remove string literals that look like they should be localized, but are never displayed.  This makes the code less complex and saves times for the localizers.

I also added a stub for the Greek locale, because Theo has offered to localize FMSLogo into Greek.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1790 9d0050d7-00b0-40f9-871c-8e02cbd5794f