Refactoring: remove g_PrinterIsCustomSetting.
This global variable was more trouble than it was worth.  It was a flag that had to be kept in sync with the screen dimensions and the active area.  But calculating it isn't expensive and was sometimes not even necessary.  This change replaces all necessary uses of the global variable with a function that does the calculation at the time it is used.  It also removes all unnecessary uses.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1852 9d0050d7-00b0-40f9-871c-8e02cbd5794f