Improve the experience of what happens when the editor contains a syntax error.  I removed the MB_QUESTIONICON because the MSDN says that this only causes confusion.  I also replaced the OK/Cancel with YES/NO, since "Cancel" usually means "go back to where I came from", not "throw away all of my data".
I also replace the title of the messagebox when the editor can't be saved.  I didn't like that FAILED was in all-caps (it's not something I want to shout at children).  Instead, the titlebar contains the name of the operation that failed, and the text better explains what went wrong.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1876 9d0050d7-00b0-40f9-871c-8e02cbd5794f