Add some tests for SETACTIVEAREA and TURTLE.
I18N fix: Replace all instances where we expect "TRUE and "FALSE to use the values that are correct with whatever locale FMSLogo was built with.

This enables me to test the localized versions of FMSLogo with the same test code.  It also enables me to find bugs in library functions when they return the incorrect form of truth and falsity (that is, if they have hard-coded a word like "true or "false.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1891 9d0050d7-00b0-40f9-871c-8e02cbd5794f