I18N fix: Make is so that "true and "vrai are not equal on the French builds.
Brian Harvey called that perverse and convinced me that the should be not equal.  There is nothing about "true and "false as words that make them true and false.  Instead, there are some function, such as IF, which assign a special meaning to some words and not others.  This is the way untyped languages like Logo are designed to work, so I should not change that.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1892 9d0050d7-00b0-40f9-871c-8e02cbd5794f