Fix an internationalization bug where, if the text of the buttons was too long, the buttons would get cut off.  This happened on the French version.
Now the dialog box dynamically resizes to fit the natural width of the buttons.

I also reduced the amount of padding between the text and the button edges, which seems to have eliminated the need for resizing the dialog box.  But I'm keeping the resizing in place in case there's another translation that is even longer than the French translation.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2105 9d0050d7-00b0-40f9-871c-8e02cbd5794f