Fix bug #1652924; FMSLogo can (once again) start up if the window was made very small before it was last closed.
This fixes the logic of how the splitter is placed to always be from 10% to 90% of the main window size.  This fixes a bug where corrupt registry setting would prevent FMSLogo from starting (and if it did start, the commander or the screen might have been completely obscured).  This bugfix is some defensive programming so that, when the program gets into a bad state, it tends toward a good state.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2140 9d0050d7-00b0-40f9-871c-8e02cbd5794f