Fix bug #1327723; EDIT now finds the editor that belongs to the same instance of FMSLogo, instead of finding it by window caption.
The fix was to extend the qlist class such that arbitrary elements could be removed.  Then, whenever an editor window is created, the TWindow* gets inserted into a list of editors in TMainFrame.  Whenever an editor is destroyed, it gets removed from the list.  When EDIT is run, if the list is not empty, some element within the list is made visible.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2224 9d0050d7-00b0-40f9-871c-8e02cbd5794f