Further performance improvement on line drawing: Creating the HPEN to use when drawing a line is now done when the pencolor/size is set, instead of just before drawing each line.  This speeds up circle drawing by about 10% and squiral drawing by about 2%.
I also added some logic for printout out GDI and USER object leaks.  Unfortunately, so many objects are leaked in normal usage, that it's hard to tell what leaks are problematic and what leaks are one-time leaks.  As a result, this is only really useful if a test is run in a loop.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2266 9d0050d7-00b0-40f9-871c-8e02cbd5794f