Fix bug #1622721; FMSLogo no longer gets into a bad state when the clipboard becomes text.
The bug was that turtlepaint(), which is called from the WM_PAINT message handler, threw a stopping error.  This stopping error would be misinterpreted when the editor was closed, as having originated while processing the contents of the editor.

The fix is to make turtlepaint() not throw a stopping error.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2283 9d0050d7-00b0-40f9-871c-8e02cbd5794f