Fix bug #1609333; "Select Active Area" dialog box no longer loses UI elements after pressing CTRL+ALT+DELETE.
I'm not sure what the problem was, but it looked like Z-order of the sunken label that FMSLogo was using to etch a rectangular background somehow got moved ahead of some of the other UI elements.  This feels more like an OS bug than an FMSLogo bug, but I was able to work-around the problem by using a groupbox, instead of a sunken static label.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2284 9d0050d7-00b0-40f9-871c-8e02cbd5794f