Fix bug #1713912; Uninstalling FMSLogo as a limited user now unregisters the file association so that double-clicking on .lgo files works if FMSLogo is later installed by an administrator.
I fixed this bug by simplifying the installer script to use SHELL_CONTEXT as the registry hive, instead of checking the UserType and then executing the same logic on a HKCU or HKLM depending on its value.  Using SHELL_CONTEXT does exactly the same thing in less than half the code.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2487 9d0050d7-00b0-40f9-871c-8e02cbd5794f