Update the translation header files to include the name of the primary translator.
Alexandre wanted to include his name in the Portuguese file and it kept getting overwritten each time I ran mklocale.pl.  This showed up as a difference in CVS.  Since it seems reasonable to include the name of the primary translator in the resulting file, I have formalized it in mklocal.pl and added it to all translations.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2528 9d0050d7-00b0-40f9-871c-8e02cbd5794f