Fix bug #1551634; Integrate some MCI examples into the manual and remove the "Help -> MCI" and "Help - Using Help" menu options to remove final remanants of the defunct .hlp format.
Removing mcistrwh.hlp has a few benefits.  First .hlp files are not supported on Vista.  They are also not supported on GNU/Linux, so this could have been an impediment to porting FMSLogo to GNU/Linux.  But the most significant reason is that I was unable to locate a copyright for mcistrwh.hlp and therefore could not prove that its license was compatible with the GPL.  Distributing this file put both FMSLogo and SourceForge at unnecessary legal risk.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2641 9d0050d7-00b0-40f9-871c-8e02cbd5794f