Implement request #1955758; turtles may optionally have their own pen state.
The involved adding a PENSTATE structure to the Turtle structure and a boolean to keep track of if the PENSTATE is used.  All access to the current pen state is done through a new accessor called GetPenStateForSelectedTurtle(), which will use either the global pen state or the g_SelectedTurtle.PenState, depending on if the selected turtle has its own pen state.

Any change to the turtle now also changes the currently used brushes and updates the status window, as appropriate.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2700 9d0050d7-00b0-40f9-871c-8e02cbd5794f