Remove zooming with Ctrl+mouse wheel for FMSLogo.
Rolling the mouse wheel while holding down the Ctrl key is too easy to do by accident and there's no obvious way to fix the problem.  Since FMSLogo already provides a menu option to configure the base font, it's better to just remove this feature. Note that zooming can still be done with Ctrl+- and Ctrl+=, which is not easy to do by accident.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2892 9d0050d7-00b0-40f9-871c-8e02cbd5794f