Fix bug #2813364; SETPENCOLOR and other commands that input a color now throw recoverable bad input errors on more bad input.
This change also centralizes the code for converting a color input into a color for more consistent behavior (before, SETPIXEL and GIFSAVE used a different function and so did not accept names for colors).

This change it not a big benefit for users.  Instead, it was done to make the test pass go more smoothly.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@2958 9d0050d7-00b0-40f9-871c-8e02cbd5794f