Update all references to g_LabelFont to use FontRec and remove g_LabelFont.
g_LabelFont was a vestige from when I thought I was going to rewrite the language engine to use wxWidgets as part of the wxWidgets port.  Instead, I'm only rewriting the UI (as a first step), so it's better to us the real label font which the language engine uses.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3164 9d0050d7-00b0-40f9-871c-8e02cbd5794f