Add support for the COMBOBOX* commands.
Unfortunately, wxComboBox does not handle CBS_SIMPLE Windows comboboxes well.  In particular, there was no way to set CBS_DISABLENOSCROLL, which means that its height would vary with the number of elements within it, which makes it incompatable with the MSWLogo behavior.

To work around this, the CLogoComboBox is written as a thin wrapper around a native COMBOBOX window, with only the functionality from wxComboBox that is necessary to support the MSWLogo commands.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3221 9d0050d7-00b0-40f9-871c-8e02cbd5794f