Fix bug #1895669; which was a race condition between NETACCEPTON/NETCONNECTON and NETWORKSHUTDOWN when asynchronous DNS was used. The problem is that the HOSTENT structure which NETACCEPTON/NETCONNECTON allocates gets freed in NETWORKSHUTDOWN. Although this calls WSACleanup, that is apparently not sufficient to stop any asynchronous calls in progress, which means that the call to WSAAsyncGetHostByName() might still attempt to write into the memory after it is freed.
This fix uses a static buffer for HOSTENT, instead of dynamically allocated memory. This is an easy fix that side-steps the problem by simply not freeing the memory which WSAAsyncGetHostByName() might want to write into. It's not an ideal fix, though, since it adds 8K of static memory, even for sessions which don't use the networking API (which is nearly all sessions).


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3251 9d0050d7-00b0-40f9-871c-8e02cbd5794f