Fix bug #1568629; FMSLogo no longer leaks memory when its internal garbage collection stack is exhausted.
The fix was to eliminate the need for the stack by using the NODE structures that we are going to free as bookkeeping space.  By the contruction of gc().

Some additional work is needed to remove the now-defunct gcstack, but since this requires changing all implementations (OWL, screensaver, and wxWidgets), it will be done in a follow-up changelist.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3256 9d0050d7-00b0-40f9-871c-8e02cbd5794f