Fix bug #1471162 and #3150920; FMSLogo no longer leaks memory when calling procedures that have parameters with default values.
The bugs were caused by inadvertant reference counting that was done when restoring "register" nodes that were not intended to be reference counted.

The fix was to add special save/restore macros which won't change the reference count of the variables and use them for the unreferenced registers.  I was also able to remove some calls that I added that set the variables to NIL to prevent crashes.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3262 9d0050d7-00b0-40f9-871c-8e02cbd5794f