Rewrite the stack logic within the evaluator so that more than two registers can be pushed to the same stack frame.  This reduces the amount of code that needs to execute in order to manage the memory of the stack and yields a 30% speedup in the evaluation.
Unfortunately, this change also causes some memory corruption in the OWL version of FMSLogo when running the non-tail recursive benchmark test.  I haven't figured out what causes this, but if I can, I'll fix it in a future revision.  If not, I'll remove the OWL project and only redistribute the wxWidgets project.


git-svn-id: https://svn.code.sf.net/p/fmslogo/source@3517 9d0050d7-00b0-40f9-871c-8e02cbd5794f